Refactor: move recursive & non-recursive resolvers into new modules
Since these are basically independent (the recursive resolver uses the
non-recursive resolver as a black box, and the non-recursive resolver
doesn't use the recursive resolver at all), and I plan to add
a *third* resolver (a forwarding one), it makes sense to put them in
separate modules.
